move login fetch request to APIAdapter
