+++ frontEnd/src/components/Home/LoginButton.tsx
@@ -1,39 +1,20 @@
 import React, { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import validateUsername from "./usernameValidation";
 import loginErrorHandler from "./loginErrorHandler";
-import { apiRequestPaths, webAppPaths } from "../../paths";
+import { webAppPaths } from "../../paths";
 import LoadingSpinner from "../shared/LoadingSpinner";
-
-async function login(username: string) {
-  console.log(`Logging in ${username}`);
-  const error = validateUsername(username);
-  if (error) {
-    throw new Error(error);
-  }
-  const response = await fetch(apiRequestPaths.login, {
-    method: "POST",
-    credentials: "include",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ username }),
-  });
-
-  if (!response.ok) {
-    throw new Error(`Login Failed - server response ${response.status}`);
-  }
-} // THIS REQUEST MUST GO IN API ADAPTER
+import APIAdapter from "../adapters/APIAdapter";
 
 export default function LoginButton() {
   const navigate = useNavigate();
   const [username, setUsername] = React.useState("");
   const [isLoading, setIsLoading] = React.useState(true);
+  const request = new APIAdapter();
 
   const handleLogin = async () => {
     setIsLoading(true);
     try {
-      await login(username);
+      await request.login(username);
       navigate(webAppPaths.teams);
     } catch (error) {
       loginErrorHandler(error);
